+++ libft/ft_toupper.c
@@ -3,18 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_toupper.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rfumeron <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: mrakhman <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2018/04/16 13:57:18 by rfumeron          #+#    #+#             */
-/*   Updated: 2018/04/18 15:14:48 by rfumeron         ###   ########.fr       */
+/*   Created: 2017/11/29 21:52:10 by mrakhman          #+#    #+#             */
+/*   Updated: 2017/11/29 21:59:41 by mrakhman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
-#include "libft.h"
 
 int	ft_toupper(int c)
 {
 	if (c >= 'a' && c <= 'z')
-		return ((int)(c - 32));
+		return (c - 32);
 	return (c);
 }
